Abstracted away cohort service stub and helper functions
